feat: production hardening — 14-task security and reliability overhaul
Phase 1 (Critical):
- Deposit monitor: savepoint-based rollback prevents batch failure on duplicate tx_hash
- Wallet RPC: async retry with exponential backoff (tenacity AsyncRetrying)
- Admin UI: CSRF token on logout form, local Tailwind CSS, cookie cleanup on auth redirect

Phase 2 (High):
- Rate limiter: off-by-one fix (> → >=) for strict limit enforcement
- Database indexes: 5 new indexes on Transaction, HubRoute, PendingWithdrawal
- Balance repo: withdraw() method tracks total_withdrawn atomically
- Failed auth: public check_failed_auth/record_failed_auth on RateLimiter (≤5/min)
- Stealth cache: immutable update via dataclasses.replace()

Phase 3 (Medium):
- Network types: Dict[str, any] → Dict[str, Any]
- Database: drop_tables() guarded to dev/test only, removed dead Database class
- Config: SQL_ECHO rejected in production via model_validator
- Privacy/metrics: type annotation fixes
- API: min_length=8 on idempotency keys, Retry-After header on 429 responses

Tests: 1178 passing, 89.45% coverage. 9 new test files, 5 updated.

+++ sthrip/privacy.py
@@ -215,7 +215,7 @@ class TransactionScheduler:
             time_since_last = (now - self._last_tx_time).total_seconds()
             min_interval = self._min_interval_minutes * 60
             if time_since_last < min_interval:
-                delay = max(delay, min_interval - time_since_last)
+                delay = int(max(delay, min_interval - time_since_last))
         
         broadcast_time = now + timedelta(seconds=delay)
         
@@ -303,7 +303,7 @@ def calculate_privacy_score(
     score += min(mixin * 2, 30)
     
     # Timing variance (max 25)
-    score += min(timing_variance * 100, 25)
+    score += int(min(timing_variance * 100, 25))
     
     # No address reuse (20 points)
     if not address_reuse:
